chore(state): Change how state is checked
Changing from NgRx selectors passed from parent to child to child (and so on) to directly referencing selector in child component.
