+++ src/app/components/body/body.component.html
@@ -3,23 +3,11 @@
   @if (importComponentState() === 'open') {
     <dct-import class="data md:grid z-30 bg-base-100" />
   }
-  <dct-data
-    #main
-    role="main"
-    class="data md:grid"
-    [hasApiKey]="hasApiKey()"
-    [variablesWithCrossTabMetadata]="variablesWithCrossTabMetadata()"
-    [isFetching]="crossTabIsFetching()"
-    [variables]="filteredByGroupID()"
-    [variablesInCrossTabSelection]="variablesInCrossTab()"
-    [crossTabValuesFetched]="crossTabValuesFetched()"
-    [selectedVariables]="selectedVariables()"
-    [weights]="allWeights()"
-    [groups]="groups()"
-    [openVariable]="openVariable()"
-    [groupChanged]="groupIDChanged()"
-    [categoriesInvalid]="categoriesMissing()"
-  />
+  <dct-data #main role="main" class="data md:grid" />
 } @else {
-  <dct-cross-tabulation id="main" role="main" class="cross-tab p-5 overflow-auto" />
+  <dct-cross-tabulation
+    id="main"
+    role="main"
+    class="cross-tab p-5 overflow-auto"
+  />
 }
