+++ src/app/components/body/variables/data/data.component.html
@@ -4,7 +4,6 @@
   [crossTabValuesFetched]="crossTabValuesFetched()"
   [groupChanged]="groupChanged()"
   [groups]="groups()"
-  [hasApiKey]="hasApiKey()"
   [isFetching]="isFetching()"
   [openVariable]="openVariable()"
   [selectedVariables]="selectedVariables()"
